feat: fix bug that will cause adding duplicate dynamic to database

+++ app/models/user_dynamic.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer
+from sqlalchemy import Column, Integer, BIGINT
 
 from app.models.base import Base
 
@@ -9,8 +9,8 @@ class UserDynamic(Base):
 
     # 表的结构:
     # 动态id
-    dynamic_id = Column(Integer, primary_key=True)
+    dynamic_id = Column(BIGINT, primary_key=True)
     type_id = Column(Integer)
-    oid = Column(Integer)
+    oid = Column(BIGINT)
     # 爬取状态
     status = Column(Integer)
